==> paper_methods/Reddit_analysis/BRM-Processes/0-environment-setup.ipynb <==
# Setting up Project Environment

### Required Packages
Should I have set up a docker instance? Yeah, probably. But for clarity, here are the specific packaged I used for this project.

!pip3 install -U torch
!pip3 install -U transformers
!pip3 install -U sklearn
!pip3 install -U

### Environment folder set-up

import os

os.system('mkdir {}'.format('data'))
os.system('mkdir {}'.format('data/scrape_data'))
os.system('mkdir {}'.format('data/vecs'))
os.system('mkdir {}'.format('data/summaries'))
