==> paper_methods/Reddit_analysis/BRM-Processes/1-Data-Collection.ipynb <==
# Data Collection: Reddit Scraping using Praw API

import sys
sys.path.append('..')

import pandas as pd
import numpy as np
from webscrapers.reddit.reddit_bot import *

### Searching for Posts Concerning a Given Topic

We start off by selecting a keyword we're interested in and designating which subreddits we want data from.

key_word = 'women'
subreddits = ['MensLib', 'MensRights', 'Feminism']

subms = []
for subreddit in subreddits:
    sub = bot.subreddit(subreddit)
    subms += [submission.id for submission in list(sub.search(key_word, sort="comments", time_filter="month", limit=3))]

subms = [flatten_comment_tree(bot.submission(id=_id)) for _id in subms]
subms = pd.concat(subms, ignore_index=True)

And we then save the data to its appropriate folder.

subms.to_csv(
    "data/scrape_data/reddit-data.tsv",
    index=False,
    encoding='utf-8',
    sep='\t'
)
